week7/requirements.txt
Flask
flasgger
pymongo
python-dotenv
PyYAML